+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 (Add entries below in the order of server, console, cli, docs, others)
 
 - server: fix GraphQL type for single-row returning functions (close #7109)
+- console: add support for creation of indexes for Postgres data sources
 
 ## v2.0.6