+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - server: fix GraphQL type for single-row returning functions (close #7109)
 - console: add support for creation of indexes for Postgres data sources
 - console: allow same named queries and unnamed queries on allowlist file upload
+- console: support computed fields in permission builder
 
 ## v2.0.6