+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 (Add entries below in the order of server, console, cli, docs, others)
 
 
+## v2.0.7
+
 - server: fix v2 -> v1 downgrade bug when cron triggers exist
 - server: add index on the `event_id` column of the `hdb_cron_event_invocation_logs` table
 - server: fix GraphQL type for remote relationship field (close #7284)